requirements.txt
langchain_openai
langchain
python-dotenv
langchain_community
chromadb
pymupdf